Fixed not executing user_data in python example
AWS EC2 instances don't execute user_data if the first line is not a shebang. Because """ in python preserves line breaks, the leading line break must be either stripped like I added with lstrip() or the shebang of the user_data python string must start right after the triple quote.